refactor: bom generation
