+++ pkg/npm/npm.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"io"
 	"path/filepath"
-	"runtime"
 	"strings"
 
 	"github.com/SAP/jenkins-library/pkg/command"
@@ -15,27 +14,6 @@ import (
 	"github.com/SAP/jenkins-library/pkg/piperutils"
 	"github.com/SAP/jenkins-library/pkg/versioning"
 )
-
-const (
-	npmBomFilename                 = "bom-npm.xml"
-	tempBomFilename                = "bom-npm.json"
-	cycloneDxNpmPackageVersion     = "@cyclonedx/cyclonedx-npm@1.11.0"
-	cycloneDxBomPackageVersion     = "@cyclonedx/bom@^3.10.6"
-	cdxgenPackageVersion           = "@cyclonedx/cdxgen@^11.3.2"
-	cycloneDxCliVersion            = "v0.27.2"
-	cycloneDxNpmInstallationFolder = "./tmp" // This folder is also added to npmignore in publish.go.Any changes to this folder needs a change in publish.go publish()
-	cycloneDxSchemaVersion         = "1.4"
-)
-
-var cycloneDxCliUrl = map[struct{ os, arch string }]string{
-	{"windows", "amd64"}: "https://github.com/CycloneDX/cyclonedx-cli/releases/download/%s/cyclonedx-win-x64.exe",
-	{"windows", "arm64"}: "https://github.com/CycloneDX/cyclonedx-cli/releases/download/%s/cyclonedx-win-arm64.exe",
-	{"darwin", "amd64"}:  "https://github.com/CycloneDX/cyclonedx-cli/releases/download/%s/cyclonedx-osx-x64",
-	{"darwin", "arm64"}:  "https://github.com/CycloneDX/cyclonedx-cli/releases/download/%s/cyclonedx-osx-arm64",
-	{"linux", "amd64"}:   "https://github.com/CycloneDX/cyclonedx-cli/releases/download/%s/cyclonedx-linux-x64",
-	{"linux", "arm64"}:   "https://github.com/CycloneDX/cyclonedx-cli/releases/download/%s/cyclonedx-linux-arm64",
-	{"linux", "arm"}:     "https://github.com/CycloneDX/cyclonedx-cli/releases/download/%s/cyclonedx-linux-arm",
-}
 
 // Execute struct holds utils to enable mocking and common parameters
 type Execute struct {
@@ -377,124 +355,5 @@ func (exec *Execute) install(packageJSON string) error {
 	if err != nil {
 		return fmt.Errorf("failed to change back into original directory: %w", err)
 	}
-	return nil
-}
-
-// CreateBOM generates BOM file using CycloneDX from all package.json files
-func (exec *Execute) CreateBOM(packageJSONFiles []string) error {
-	// Detect package manager to determine which BOM generator to use
-	pm, err := exec.detectPackageManager()
-	if err != nil {
-		return err
-	}
-
-	// Install cyclonedx-npm in a new folder (to avoid extraneous errors) and generate BOM
-	cycloneDxNpmInstallParams := []string{"install", "--no-save", cycloneDxNpmPackageVersion, "--prefix", cycloneDxNpmInstallationFolder}
-	cycloneDxNpmRunParams := []string{"--output-format", "XML", "--spec-version", cycloneDxSchemaVersion, "--omit", "dev", "--output-file"}
-
-	// Install cyclonedx/bom with --nosave and generate BOM.
-	cycloneDxBomInstallParams := []string{"install", cycloneDxBomPackageVersion, "--no-save"}
-	cycloneDxBomRunParams := []string{"cyclonedx-bom", "--output"}
-
-	// Attempt#1, generate BOM via cyclonedx-npm
-	// Check for pnpm first since it uses cdxgen
-	if pm != nil && pm.Name == "pnpm" {
-		// Download cyclonedx-cli
-		osArch := struct{ os, arch string }{runtime.GOOS, runtime.GOARCH}
-		var cliPath string
-		var err error
-		if urlTemplate, ok := cycloneDxCliUrl[osArch]; ok {
-			url := fmt.Sprintf(urlTemplate, cycloneDxCliVersion)
-			cliPath, err = piperhttp.DownloadExecutable("", exec.Utils.GetFileUtils(), exec.Utils.GetDownloadUtils(), url)
-			if err != nil {
-				return fmt.Errorf("failed to download cyclonedx-cli: %w", err)
-			}
-		} else {
-			return fmt.Errorf("cyclonedx-cli not available for OS/architecture combination: %s/%s", runtime.GOOS, runtime.GOARCH)
-		}
-
-		// Install cdxgen for pnpm projects
-		execRunner := exec.Utils.GetExecRunner()
-		err = execRunner.RunExecutable("npm", "install", cdxgenPackageVersion, "--prefix", cycloneDxNpmInstallationFolder)
-		if err != nil {
-			return fmt.Errorf("failed to install cdxgen: %w", err)
-		}
-
-		for _, packageJSONFile := range packageJSONFiles {
-			path := filepath.Dir(packageJSONFile)
-			jsonBomPath := filepath.Join(path, tempBomFilename)
-			xmlBomPath := filepath.Join(path, npmBomFilename)
-
-			// Generate JSON SBOM with cdxgen
-			cdxgenExecutable := cycloneDxNpmInstallationFolder + "/node_modules/.bin/cdxgen"
-			params := []string{
-				"-r",
-				"-o", jsonBomPath,
-				"--spec-version", cycloneDxSchemaVersion,
-			}
-			err := execRunner.RunExecutable(cdxgenExecutable, params...)
-			if err != nil {
-				return fmt.Errorf("failed to generate CycloneDX BOM with cdxgen: %w", err)
-			}
-
-			log.Entry().Infof("Generated CycloneDX BOM in JSON format at %s", jsonBomPath)
-
-			// Convert JSON to XML using cyclonedx-cli
-			err = execRunner.RunExecutable(cliPath, "convert", "--input-file", jsonBomPath, "--output-format", "xml", "--output-file", xmlBomPath)
-			if err != nil {
-				return fmt.Errorf("failed to convert BOM to XML format: %w", err)
-			}
-
-			log.Entry().Infof("Converted CycloneDX BOM to XML format at %s", xmlBomPath)
-		}
-	} else {
-		// For non-pnpm projects, use existing CycloneDX approach
-		err := exec.createBOMWithParams(cycloneDxNpmInstallParams, cycloneDxNpmRunParams, packageJSONFiles, false)
-		if err != nil {
-			log.Entry().Infof("Failed to generate BOM CycloneDX BOM with cyclonedx-npm ,fallback to cyclonedx/bom")
-
-			// Attempt #2, generate BOM via cyclonedx/bom@^3.10.6
-			err = exec.createBOMWithParams(cycloneDxBomInstallParams, cycloneDxBomRunParams, packageJSONFiles, true)
-			if err != nil {
-				log.Entry().Infof("Failed to generate BOM CycloneDX BOM with fallback package cyclonedx/bom ")
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-// Facilitates BOM generation with different packages
-func (exec *Execute) createBOMWithParams(packageInstallParams []string, packageRunParams []string, packageJSONFiles []string, fallback bool) error {
-	execRunner := exec.Utils.GetExecRunner()
-
-	// Install package
-	err := execRunner.RunExecutable("npm", packageInstallParams...)
-	if err != nil {
-		return fmt.Errorf("failed to install CycloneDX BOM %w", err)
-	}
-
-	// Run package for all package JSON files
-	if len(packageJSONFiles) > 0 {
-		for _, packageJSONFile := range packageJSONFiles {
-			path := filepath.Dir(packageJSONFile)
-			executable := "npx"
-			params := append(packageRunParams, filepath.Join(path, npmBomFilename))
-
-			// Below code needed as to adjust according to needs of cyclonedx-npm and fallback cyclonedx/bom@^3.10.6
-			if !fallback {
-				params = append(params, packageJSONFile)
-				executable = cycloneDxNpmInstallationFolder + "/node_modules/.bin/cyclonedx-npm"
-			} else {
-				params = append(params, path)
-			}
-
-			err := execRunner.RunExecutable(executable, params...)
-			if err != nil {
-				return fmt.Errorf("failed to generate CycloneDX BOM :%w", err)
-			}
-		}
-	}
-
 	return nil
 }
